added more new event handling regarding input elements
The bet buttons are now only active before a game starts. Bet can be changed up to the moment player presses the "Throw" button. After the first round starts the buttons are disabled.

Additionally, the "Throw" button is only visible after the first bet value is chosen, and disappears when a winner has been chosen.