changed how to handles betPanel visibility
betpanel visibility now reacts to the state of  specific variables.

if there are enough funds to make a bet an the bet has not been looked in (the bet has been chosen but not registered). then the betpanel is visible and bet process is initiated.

if a bet has been locked in (bet != 0 and IsLockedIn) then the betpanel is visible

if neither of these are true then the bet panel is not visible